refactor(Team): 패키지 이동
- Team이 piece 이외에도 범용적으로 이동되면서, 패키지 이동 실시
