+++ src/main/java/chess/domain/piece/Bishop.java
@@ -1,5 +1,6 @@
 package chess.domain.piece;
 
+import chess.domain.Team;
 import chess.domain.movement.MovementRule;
 import chess.domain.movement.continuous.NorthEastMovement;
 import chess.domain.movement.continuous.NorthWestMovement;
