+++ src/main/java/chess/domain/piece/Pawn.java
@@ -1,5 +1,6 @@
 package chess.domain.piece;
 
+import chess.domain.Team;
 import chess.domain.movement.BlackPawnFirstMovement;
 import chess.domain.movement.MovementRule;
 import chess.domain.movement.WhitePawnFirstMovement;
